+++ src/twg2/parser/fragment/TextToken.java
@@ -13,4 +13,10 @@ public interface TextToken<S, T> {
 	@Override
 	public String toString();
 
+	@Override
+	public int hashCode();
+
+	@Override
+	public boolean equals(Object obj);
+
 }